Make it easier to run test harness with a single view
