+++ crates/viewer/re_view_dataframe/Cargo.toml
@@ -20,7 +20,7 @@ all-features = true
 
 [features]
 default = []
-testing = ["re_viewer_context/testing", "re_viewport_blueprint/testing"]
+testing = ["re_viewer_context/testing"]
 
 [dependencies]
 re_chunk_store.workspace = true
@@ -49,5 +49,4 @@ itertools.workspace = true
 re_chunk_store.workspace = true
 re_component_ui.workspace = true
 re_viewer_context = { workspace = true, features = ["testing"] }
-re_viewport.workspace = true
-re_viewport_blueprint = { workspace = true, features = ["testing"] }
+re_viewport = { workspace = true, features = ["testing"] }
