+++ crates/viewer/re_view_graph/tests/basic.rs
@@ -8,7 +8,8 @@ use re_types::archetypes;
 use re_view_graph::GraphView;
 use re_viewer_context::test_context::HarnessExt as _;
 use re_viewer_context::{ViewClass as _, test_context::TestContext};
-use re_viewport_blueprint::{ViewBlueprint, test_context_ext::TestContextExt as _};
+use re_viewport::test_context_ext::TestContextExt as _;
+use re_viewport_blueprint::ViewBlueprint;
 
 #[test]
 pub fn coincident_nodes() {
@@ -143,31 +144,8 @@ fn run_graph_view_and_save_snapshot(test_context: &mut TestContext, name: &str, 
         .setup_kittest_for_rendering()
         .with_size(size)
         .with_max_steps(256) // Give it some time to settle the graph.
-        .build_ui(|ui| {
-            test_context.run(&ui.ctx().clone(), |ctx| {
-                let view_class = ctx
-                    .view_class_registry()
-                    .get_class_or_log_error(GraphView::identifier());
-
-                let view_blueprint = ViewBlueprint::try_from_db(
-                    view_id,
-                    ctx.store_context.blueprint,
-                    ctx.blueprint_query,
-                )
-                .expect("we just created that view");
-
-                let mut view_states = test_context.view_states.lock();
-                let view_state = view_states.get_mut_or_create(view_id, view_class);
-
-                let (view_query, system_execution_output) =
-                    re_viewport::execute_systems_for_view(ctx, &view_blueprint, view_state);
-
-                view_class
-                    .ui(ctx, ui, view_state, &view_query, system_execution_output)
-                    .expect("failed to run graph view ui");
-            });
-
-            test_context.handle_system_commands();
+        .build(|ctx| {
+            test_context.run_with_single_view(ctx, view_id);
         });
 
     harness.run();
