+++ crates/viewer/re_view_graph/tests/snapshots/coincident_nodes.png
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:42c8ccb7f1f2d41db5c9db12e8c338d681281b474672944e5dec1b522cae9e92
-size 848
+oid sha256:09db7e1dd99757468b47a57cd37e2597178338b71edcfa7b463b8c547fddb738
+size 833
