+++ crates/viewer/re_view_graph/tests/snapshots/multi_graphs.png
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:e6a67770a64097492c12813bc9dce6426dd88b98740c4a4cdec6b5450f6acbe7
-size 6770
+oid sha256:906efa06d8e2903436764a99b79ced9240057e0feef99abb6a68e1ca55c526fd
+size 6723
