+++ crates/viewer/re_view_graph/tests/snapshots/self_and_multi_edges.png
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:59f399c7410ff3e28511ab2ee7a690cc35be1c97bed99c6f31209ba968a1e6b9
-size 20082
+oid sha256:91d73e5a837e14811f2c9926d921bb583729b374343b51432a096b51931e47fe
+size 19031
