+++ crates/viewer/re_view_spatial/tests/transform_clamping.rs
@@ -1,10 +1,9 @@
 use re_chunk_store::RowId;
 use re_log_types::TimePoint;
-use re_view_spatial::SpatialView3D;
 use re_viewer_context::test_context::TestContext;
 use re_viewer_context::{RecommendedView, ViewClass as _, ViewId};
+use re_viewport::test_context_ext::TestContextExt as _;
 use re_viewport_blueprint::ViewBlueprint;
-use re_viewport_blueprint::test_context_ext::TestContextExt as _;
 
 #[test]
 pub fn test_transform_clamping() {
@@ -203,33 +202,7 @@ fn run_view_ui_and_save_snapshot(
             .setup_kittest_for_rendering()
             .with_size(size)
             .build(|ctx| {
-                re_ui::apply_style_and_install_loaders(ctx);
-
-                egui::CentralPanel::default().show(ctx, |ui| {
-                    test_context.run(ctx, |ctx| {
-                        let view_class = ctx
-                            .view_class_registry()
-                            .get_class_or_log_error(SpatialView3D::identifier());
-
-                        let view_blueprint = ViewBlueprint::try_from_db(
-                            view_id,
-                            ctx.store_context.blueprint,
-                            ctx.blueprint_query,
-                        )
-                        .expect("we just created that view");
-
-                        let mut view_states = test_context.view_states.lock();
-
-                        let view_state = view_states.get_mut_or_create(view_id, view_class);
-                        let (view_query, system_execution_output) =
-                            re_viewport::execute_systems_for_view(ctx, &view_blueprint, view_state);
-                        view_class
-                            .ui(ctx, ui, view_state, &view_query, system_execution_output)
-                            .expect("failed to run view ui");
-                    });
-
-                    test_context.handle_system_commands();
-                });
+                test_context.run_with_single_view(ctx, view_id);
             });
 
         {
