+++ crates/viewer/re_viewport/Cargo.toml
@@ -18,6 +18,10 @@ workspace = true
 [package.metadata.docs.rs]
 all-features = true
 
+[features]
+## Enable for testing utilities.
+testing = ["re_viewer_context/testing"]
+
 [dependencies]
 re_context_menu.workspace = true
 re_entity_db.workspace = true
